[4] Documentar deposits USDC (#17)
* docs

* refactor: moved docs

* refactor: moved docs

* fix: redundant types
